C interface divided into separate files

+++ include/gotchangpdf/c_integer_object.h
@@ -9,7 +9,7 @@ extern "C"
 {
 #endif
 
-	GOTCHANG_PDF_API long long CALLING_CONVENTION IntegerObject_Value(IntegerObjectHandle);
+	GOTCHANG_PDF_API int CALLING_CONVENTION IntegerObject_Value(IntegerObjectHandle);
 	GOTCHANG_PDF_API void CALLING_CONVENTION IntegerObject_Release(IntegerObjectHandle);
 
 #ifdef __cplusplus
